#8 [refact] 회원가입 시 user도메인 만드는 메소드 User파일에서 SignUpRequest에서 만드는 것으로 수정
